Create pekingnese_generator.py
This code takes user input, converts it to a list of characters, and then iterates through each character. If the character is alphanumeric (i.e. not a punctuation mark), it adds "儿" to the end of the character before appending it to the result string. If the character is a punctuation mark, it simply appends it to the result string without adding anything. Finally, it prints out the result string